Refactor Minmod work into minmod_impl
This improves consistency with WENO limiters
